requirements.txt
flask==3.1.0
google-auth==2.38.0
google-api-python-client==2.163.0
gunicorn==23.0.0
